Extract out BulkIndexer
Isolate this logic to it's own class.
