+++ lib/elasticsearch/extensions/documents.rb
@@ -5,6 +5,7 @@ require "elasticsearch/extensions/documents/version"
 require "elasticsearch/extensions/documents/document"
 require "elasticsearch/extensions/documents/index"
 require "elasticsearch/extensions/documents/indexer"
+require "elasticsearch/extensions/documents/bulk_indexer"
 require "elasticsearch/extensions/documents/queryable"
 require "elasticsearch/extensions/documents/utils"
