+++ lib/elasticsearch/extensions/documents/indexer.rb
@@ -18,10 +18,10 @@ module Elasticsearch
         end
 
         def create_index
-          client.indices.create(index: config.index_name, body: create_index_body) unless index_exists?
+          client.indices.create(index: config.index_name, body: index_definition) unless index_exists?
         end
 
-        def create_index_body
+        def index_definition
           {}.tap do |body|
             body[:settings] = config.settings if config.settings
             body[:mappings] = config.mappings if config.mappings
@@ -35,25 +35,6 @@ module Elasticsearch
           create_index
 
           block.call(self) if block_given?
-        end
-
-        def bulk_index(documents)
-          client.bulk body: bulk_index_operations(documents)
-        end
-
-        def bulk_index_operations(documents)
-          documents.collect { |document| bulk_index_operation_hash(document) }
-        end
-
-        def bulk_index_operation_hash(document)
-          {
-            index: {
-              _index: config.index_name,
-              _type:  document.class.type,
-              _id:    document.id,
-              data:   document.as_hash,
-            }
-          }
         end
 
       end
